11. Commit - 11.08.2023
Added handleAddInput Function: The handleAddInput function is introduced to allow adding new input fields when clicking the "+" icon, ensuring that a maximum of 10 inputs can be added.

Added "+"" Icon to Last Input: In the rendering loop, a "+" icon is now added next to the last input field, allowing the user to add new inputs dynamically.

Added Reload on Save: After successful data update, the handleSave function triggers a page reload using window.location.reload(), ensuring that the updated data is reloaded and displayed.

Improved fetchAndSetNavbarData: The fetchAndSetNavbarData function now fetches and sets the navbar data on component mount, ensuring that the data is loaded when the component is first rendered.

Adjusted Padding Based on Data Length: The showPadding variable dynamically adjusts the left padding of the content based on the number of displayed items. If there are 6 or fewer items, the padding is set to "20%"; otherwise, it is set to "10%".

Refined Code Structure: Unnecessary or unused functions and code segments have been removed to streamline the component and improve its readability and maintainability.

These changes collectively enhance the functionality of the component by allowing dynamic addition of input fields, triggering a page reload after saving, and optimizing the code structure.