39. Commit - 23.08.2023
In the sidebar image enlargement feature, we were facing an issue where only the same image was consistently being zoomed in. To address this, we introduced a variable named "enlargedImages." This variable keeps track of whether images have been zoomed in previously. During the image enlargement process, we utilized this variable to check if the image had been enlarged before.

Furthermore, when an image was zoomed in, there was an issue where clicking on the image was not responsive due to the "image-overlay" covering the image. To resolve this, we adjusted the "pointerEvents" property accordingly. This allowed us to make the image clickable by appropriately configuring the "pointerEvents" property when the image was zoomed in.