50. Commit - 1.09.2023
Fix Admin Login Error and Enhance Security

This commit addresses several issues and improvements:

1. Fixed the Admin Login error by moving the compare method from the server to the client.
2. Improved security measures by migrating the compare method to the client-side.
3. Resolved the saved_version error by adding default color options since colors cannot be empty in the MySQL database.
4. The Insert button now works seamlessly with saved_versions.

These changes enhance the reliability and security of the application while providing a more user-friendly experience. The addition of default color options ensures data consistency in the MySQL database, and moving the compare method to the client-side improves the overall performance of the application.